fix(Toggle): [B2BDS-190] only emit change event when new value selected
refactor(Toggle): [B2BDS-190] revert change as it is not needed

this is already checked in the toggle button, no need to check again in group.

fix(Toggle): [B2BDS-190] only emit change event when new value selected

refactor(Toggle): [B2BDS-190] revert change as it is not needed

this is already checked in the toggle button, no need to check again in group.